requirements.txt
dvc
mlflow
scikit-learn
pandas
numpy
pyyaml
joblib
